working AntBase
